# Conflicts:
#	backend/src/main/java/tuneandmanner/wiselydiarybackend/cartoon/domain/entity/DiarySummary.java
#	backend/src/main/java/tuneandmanner/wiselydiarybackend/cartoon/service/CartoonService.java

+++ backend/src/main/java/tuneandmanner/wiselydiarybackend/cartoon/dto/request/SaveCartoonRequest.java
@@ -11,5 +11,5 @@ import lombok.Setter;
 @AllArgsConstructor
 public class SaveCartoonRequest {
     private String cartoonPath;
-    private Integer diarySummaryCode;
+    private Long diarySummaryCode;
 }
